Implement collectors
A collector is an object having the ability to collect information from
a specific provider (i.e. 'yelp' or 'google').

This patch implements a Google and a Yelp collector as well as a generic
client collector which abstracts all the mechanisms to authenticate and
retrieve the exact information we want.

Miscellaneous items:
* Update `.gitignore` file
* Update Makefile
* Update test dependencies:
  * Add `Faker`
  * Add `pytest-mock`
* Update Sphinx documentation structure
* Update the documentation
* Use the basicstrap sphinx theme for the documentation
* Add unit tests
* Fix tox.ini
* Fix coala
  * The output is broken in the `pre` version
  (see https://github.com/coala/coala/pull/4420 for more details)
* Ensure the `ci-*` targets are passing